+++ src/App.js
@@ -6,6 +6,7 @@ import styled from 'styled-components';
 import { firebase, db } from './firebase';
 import useDoc from './hooks/useDoc';
 import HomePage from './components/HomePage';
+import Choose from './components/Choose';
 
 const Wrapper = styled.div`
   border: 3px dashed black;
@@ -152,173 +153,20 @@ function Login() {
     </Wrapper>
   );
 }
-function Choose({ user }) {
-  const [newUser, setNewUser] = React.useState(null);
-  React.useEffect(() => {
-    if (newUser) {
-      db.doc(`users/${user.uid}`).update({
-        contact: {
-          email: newUser.contact.email,
-          name: newUser.contact.name,
-          phoneNumber: newUser.contact.phoneNumber,
-        },
-      });
-    }
-  }, [newUser, user.uid]);
-  const updateUser = values => {
-    const formattedPhone = String('+1').concat(
-      String(values.phoneNumber).replace(/[^\d]/g, ''),
-    );
-    setNewUser({
-      ...user,
-      contact: { ...values, phoneNumber: formattedPhone },
-    });
-  };
-  const currentUserData = useDoc(`users/${user.uid}`);
-  if (newUser || currentUserData) {
-    if (currentUserData && currentUserData.contact) {
-      return <HomePage user={currentUserData} />;
-    }
-    if (newUser && newUser.contact) {
-      return <HomePage user={newUser} />;
-    }
-  }
-  console.log(user, 'user');
-  return (
-    <>
-      <div>Hello {user.displayName} </div>
-      <div className='app'>
-        <h1>Choose Your Loved One</h1>
-        <button
-          type='button'
-          onClick={() => {
-            firebase.auth().signOut();
-          }}
-        >
-          log out
-        </button>
-
-        <Formik
-          initialValues={{ email: '', name: '', phoneNumber: '' }}
-          onSubmit={(values, { setSubmitting }) => {
-            setTimeout(() => {
-              updateUser(values);
-              setSubmitting(false);
-            }, 500);
-          }}
-          validationSchema={Yup.object().shape({
-            email: Yup.string()
-              .email()
-              .required('Required'),
-            name: Yup.string()
-              .min(2, 'Too Short!')
-              .max(50, 'Too Long!')
-              .required('Required'),
-            phoneNumber: Yup.number().required('Required'),
-          })}
-        >
-          {props => {
-            const {
-              values,
-              touched,
-              errors,
-              dirty,
-              isSubmitting,
-              handleChange,
-              handleBlur,
-              handleSubmit,
-              handleReset,
-            } = props;
-            return (
-              <form onSubmit={handleSubmit}>
-                <label htmlFor='email' style={{ display: 'block' }}>
-                  Email
-                </label>
-                <input
-                  id='email'
-                  placeholder='Enter your email'
-                  type='text'
-                  value={values.email}
-                  onChange={handleChange}
-                  onBlur={handleBlur}
-                  className={
-                    errors.email && touched.email
-                      ? 'text-input error'
-                      : 'text-input'
-                  }
-                />
-                {errors.email && touched.email && (
-                  <div className='input-feedback'>{errors.email}</div>
-                )}
-                <label htmlFor='name' style={{ display: 'block' }}>
-                  Name
-                </label>
-                <input
-                  id='name'
-                  placeholder='Enter your name'
-                  type='text'
-                  value={values.name}
-                  onChange={handleChange}
-                  onBlur={handleBlur}
-                  className={
-                    errors.name && touched.name
-                      ? 'text-input error'
-                      : 'text-input'
-                  }
-                />
-                {errors.name && touched.name && (
-                  <div className='input-feedback'>{errors.name}</div>
-                )}
-                <label htmlFor='phoneNumber' style={{ display: 'block' }}>
-                  Phone Number
-                </label>
-                <input
-                  id='phoneNumber'
-                  placeholder='Enter your phone number'
-                  type='number'
-                  value={values.phoneNumber}
-                  onChange={handleChange}
-                  onBlur={handleBlur}
-                  className={
-                    errors.phoneNumber && touched.phoneNumber
-                      ? 'text-input error'
-                      : 'text-input'
-                  }
-                />
-                {errors.phoneNumber && touched.phoneNumber && (
-                  <div className='input-feedback'>{errors.phoneNumber}</div>
-                )}
-                <button
-                  type='button'
-                  className='outline'
-                  onClick={handleReset}
-                  disabled={!dirty || isSubmitting}
-                >
-                  Reset
-                </button>
-                <button type='submit' disabled={isSubmitting}>
-                  Submit
-                </button>
-              </form>
-            );
-          }}
-        </Formik>
-      </div>
-    </>
-  );
-}
 
 const handleSubmit = async event => {
   event.preventDefault();
-
+  const userEmail = 'bobros134@bob.com';
+  const password = '123456';
   try {
     const { user } = await firebase
       .auth()
-      .createUserWithEmailAndPassword('bobross@bob.com', '123456');
-
+      .createUserWithEmailAndPassword(userEmail, password);
     user.updateProfile({ displayName: 'Bob Ross' });
+    console.log(user, '$$');
   } catch (error) {
     alert(error);
   }
 };
+
 export default App;
